% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nres}
\alias{theme_nres}
\title{Default NRES theme}
\usage{
theme_nres(base_size = 16, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family (default = sans)}
}
\value{
a ggplot2 theme object
}
\description{
Default NRES theme
}
